requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
pyparsing==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
urllib3==1.25.3
